@@ -15,8 +15,8 @@ import com.cn.speaktest.service.ExamService
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
-import javax.validation.Valid
-import javax.validation.constraints.NotBlank
+import jakarta.validation.Valid
+import jakarta.validation.constraints.NotBlank
 
 @CrossOrigin(origins = ["*"], maxAge = 3600)
 @RestController
