@@ -1,8 +1,8 @@
 package com.cn.speaktest.payload.request.auth
 
-import javax.validation.constraints.Email
-import javax.validation.constraints.NotBlank
-import javax.validation.constraints.Size
+import jakarta.validation.constraints.Email
+import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.Size
 
 data class SignupRequest(
     var fullName: String?,
