@@ -1,6 +1,6 @@
 package com.cn.speaktest.payload.request.auth
 
-import javax.validation.constraints.NotBlank
+import jakarta.validation.constraints.NotBlank
 
 data class TokenRefreshRequest (
     @field:NotBlank var refreshToken: String
