@@ -1,14 +1,15 @@
 package com.cn.speaktest.security.jwt
 
 import com.cn.speaktest.advice.Message
+import com.cn.speaktest.advice.toJson
 import org.springframework.http.HttpStatus
 import org.springframework.security.core.AuthenticationException
 import org.springframework.security.web.AuthenticationEntryPoint
 import org.springframework.stereotype.Component
 import java.io.IOException
-import javax.servlet.ServletException
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.servlet.ServletException
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 
 @Component
 class AuthEntryPointJwt : AuthenticationEntryPoint {
@@ -27,13 +28,4 @@ class AuthEntryPointJwt : AuthenticationEntryPoint {
             ).toJson()
         )
     }
-}
-
-private fun Message.toJson(): String {
-    return "{\n" +
-            "    \"status\": \"$status\",\n" +
-            "    \"code\": $code,\n" +
-            "    \"message\": \"$message\",\n" +
-            "    \"data\": $data\n" +
-            "}"
 }
