@@ -1,44 +1,52 @@
 package com.cn.speaktest.config
 
+import com.cn.speaktest.advice.CustomAccessDeniedHandler
 import com.cn.speaktest.security.jwt.AuthEntryPointJwt
 import com.cn.speaktest.security.jwt.AuthTokenFilter
 import com.cn.speaktest.security.jwt.JwtUtils
 import com.cn.speaktest.security.services.UserDetailsServiceImpl
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
 import org.springframework.security.config.http.SessionCreationPolicy
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
 import org.springframework.security.crypto.password.PasswordEncoder
+import org.springframework.security.web.SecurityFilterChain
+import org.springframework.security.web.access.AccessDeniedHandler
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
+import org.springframework.web.cors.CorsConfiguration
+import org.springframework.web.cors.CorsConfigurationSource
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource
 
 @Configuration
 @EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
+@EnableMethodSecurity(prePostEnabled = true)
 class WebSecurityConfig(
     private val jwtUtils: JwtUtils,
     private val userDetailsService: UserDetailsServiceImpl,
     private val unauthorizedHandler: AuthEntryPointJwt,
-) : WebSecurityConfigurerAdapter() {
+) {
+
+    @Autowired
+    fun registerProvider(auth: AuthenticationManagerBuilder) {
+        auth.userDetailsService(userDetailsService).passwordEncoder(BCryptPasswordEncoder())
+    }
+
+    @Bean
+    @Throws(java.lang.Exception::class)
+    fun authenticationManager(authenticationConfiguration: AuthenticationConfiguration): AuthenticationManager? {
+        return authenticationConfiguration.authenticationManager
+    }
+
     @Bean
     fun authenticationJwtTokenFilter(): AuthTokenFilter {
         return AuthTokenFilter(jwtUtils, userDetailsService)
-    }
-
-    @Throws(Exception::class)
-    public override fun configure(authenticationManagerBuilder: AuthenticationManagerBuilder) {
-        authenticationManagerBuilder.userDetailsService(userDetailsService).passwordEncoder(passwordEncoder())
-    }
-
-    @Bean
-    @Throws(Exception::class)
-    override fun authenticationManagerBean(): AuthenticationManager {
-        return super.authenticationManagerBean()
     }
 
     @Bean
@@ -46,21 +54,50 @@ class WebSecurityConfig(
         return BCryptPasswordEncoder()
     }
 
+    @Bean
+    fun accessDeniedHandler(): AccessDeniedHandler {
+        return CustomAccessDeniedHandler()
+    }
+
+    @Bean
     @Throws(Exception::class)
-    override fun configure(http: HttpSecurity) {
-        http.cors().and().csrf().disable()
+    fun filterChain(http: HttpSecurity): SecurityFilterChain? {
+        return http
+            .cors().and()
+            .csrf().disable()
+            .authorizeHttpRequests().anyRequest().permitAll().and()
             .exceptionHandling().authenticationEntryPoint(unauthorizedHandler).and()
+            .exceptionHandling().accessDeniedHandler(accessDeniedHandler()).and()
             .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
-            .authorizeRequests().antMatchers(
-                "/api/auth/**",
-                "/swagger-ui/**",
-                "/v3/api-docs/**",
-                "/swagger-ui.html",
-                "/api/docs/**"
+            .addFilterBefore(
+                AuthTokenFilter(jwtUtils, userDetailsService),
+                UsernamePasswordAuthenticationFilter::class.java
             )
-            .permitAll()
-            .anyRequest()
-            .authenticated()
-        http.addFilterBefore(authenticationJwtTokenFilter(), UsernamePasswordAuthenticationFilter::class.java)
+            .build()
+    }
+
+    @Bean
+    fun corsConfigurationSource(): CorsConfigurationSource = UrlBasedCorsConfigurationSource().also { cors ->
+        CorsConfiguration().apply {
+            allowedOrigins = listOf("*")
+            allowedMethods = listOf("POST", "PUT", "DELETE", "GET", "OPTIONS", "HEAD")
+            allowedHeaders = listOf(
+                "Authorization",
+                "Content-Type",
+                "X-Requested-With",
+                "Accept",
+                "Origin",
+                "Access-Control-Request-Method",
+                "Access-Control-Request-Headers"
+            )
+            exposedHeaders = listOf(
+                "Access-Control-Allow-Origin",
+                "Access-Control-Allow-Credentials",
+                "Authorization",
+                "Content-Disposition"
+            )
+            maxAge = 3600
+            cors.registerCorsConfiguration("/**", this)
+        }
     }
 }
